modul-41

+++ app/src/main/java/com/example/project/data/MainRepository.kt
@@ -2,6 +2,7 @@ package com.example.project.data
 
 import android.content.ContentValues
 import android.database.Cursor
+import androidx.lifecycle.LiveData
 import com.example.project.data.DAO.FilmDao
 import com.example.project.data.entity.Film
 import java.util.concurrent.Executors
@@ -14,21 +15,22 @@ class MainRepository(private val filmDao: FilmDao) {
         }
     }
 
-    fun getAllFromDB(): List<Film> {
+    fun getAllFromDB(): LiveData<List<Film>> {
         return filmDao.getCachedFilms()
     }
     //Очистка кэша
     fun clearCache(){
         Executors.newSingleThreadExecutor().execute {
-            val list = filmDao.getCachedFilms()
-            filmDao.deleteDB(list)
+            val list = filmDao.getCachedFilms().value?.toList()
+            if (list != null)
+                filmDao.deleteDB(list)
         }
     }
     //Очистить кэш от фильмов с рейтингом ниже 8.0
     fun clearInCacheBadFilms(){
         Executors.newSingleThreadExecutor().execute {
             val list = filmDao.getCachedFilmsGood(0.0, 7.99)
-            list.forEach{
+            list.value?.forEach{
                 filmDao.deleteFilm(it)
             }
         }
